+++ addon/services/swal.js
@@ -94,6 +94,10 @@ export default class SweetAlertService extends Service {
     return this.sweetAlert.getConfirmButton();
   }
 
+  getDenyButton() {
+    return this.sweetAlert.getDenyButton();
+  }
+
   getCancelButton() {
     return this.sweetAlert.getCancelButton();
   }
@@ -110,16 +114,8 @@ export default class SweetAlertService extends Service {
     this._run('showLoading');
   }
 
-  enableLoading() {
-    this._run('enableLoading');
-  }
-
   hideLoading() {
     this._run('hideLoading');
-  }
-
-  disableLoading() {
-    this._run('disableLoading');
   }
 
   isLoading() {
@@ -152,6 +148,10 @@ export default class SweetAlertService extends Service {
 
   clickConfirm() {
     this._run('clickConfirm');
+  }
+
+  clickDeny() {
+    this._run('clickDeny');
   }
 
   clickCancel() {