+++ tests/dummy/app/controllers/application.js
@@ -24,7 +24,7 @@ export default class ApplicationController extends Controller {
   @action
   loading() {
     return new Promise(resolve => {
-      this.swal.enableLoading();
+      this.swal.showLoading();
       TESTING ? next(null, resolve) : later(resolve, 2000);
     });
   }